requirements.txt
Django==4.2.7
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9